+++ packages/expo-av/android/src/main/java/expo/modules/av/video/VideoView.java
@@ -46,6 +46,7 @@ public class VideoView extends FrameLayout implements AudioEventHandler, Fullscr
 
   private PlayerData mPlayerData = null;
 
+  private ReadableArguments mLastSource;
   private ScalableType mResizeMode = ScalableType.LEFT_TOP;
   private boolean mUseNativeControls = false;
   private Boolean mOverridingUseNativeControls = null;
@@ -298,6 +299,37 @@ public class VideoView extends FrameLayout implements AudioEventHandler, Fullscr
   void setUseNativeControls(final boolean useNativeControls) {
     mUseNativeControls = useNativeControls;
     maybeUpdateMediaControllerForUseNativeControls();
+  }
+
+  private static boolean equalBundles(Bundle one, Bundle two) {
+    if((one.size() != two.size()) || !one.keySet().containsAll(two.keySet())) {
+      return false;
+    }
+
+    for (String key : one.keySet()) {
+      Object valueOne = one.get(key);
+      Object valueTwo = two.get(key);
+      if (valueOne instanceof Bundle && valueTwo instanceof Bundle) {
+        if (!equalBundles((Bundle) valueOne, (Bundle) valueTwo)) {
+          return false;
+        }
+      } else if (valueOne == null) {
+        if (valueTwo != null) {
+          return false;
+        }
+      } else if (!valueOne.equals(valueTwo)) {
+        return false;
+      }
+    }
+
+    return true;
+  }
+
+  public void setSource(final ReadableArguments source) {
+    if (mLastSource == null || !equalBundles(mLastSource.toBundle(), source.toBundle())) {
+      mLastSource = source;
+      setSource(source, null, null);
+    }
   }
 
   public void setSource(final ReadableArguments source, final ReadableArguments initialStatus, final Promise promise) {